Se muestran eventos anteriores en su propia sección; se corrige error en listado de usuarios; se cambia método get por post en eliminar; se cambia método post por get en nuevo.

+++ app/Controllers/Acceso_sistema.php
@@ -68,21 +68,29 @@ class Acceso_sistema extends BaseController
         }
     }
 
-    public function eliminar($id_acceso_sistema)
+    public function eliminar()
     {
         if ($this->session->logueado) {
 
-            // registro en bitacora
-            $acceso_sistema = $this->acceso_sistema_model->get_acceso_sistema($id_acceso_sistema);
-            $accion = "eliminó";
-            $entidad = 'acceso_sistema';
-            $valor = $acceso_sistema['id_rol'] . " " . $acceso_sistema['cod_opcion_sistema'];
-            $this->fn_sis->registro_bitacora($accion, $entidad, $valor);
+            $acceso_sistema = $this->request->getPost();
+            if ($acceso_sistema) {
+                $id_acceso_sistema = $acceso_sistema['id_acceso_sistema'];
+                $url_actual = $acceso_sistema['url_actual'];
 
-            // eliminado
-            $this->acceso_sistema_model->delete($id_acceso_sistema);
+                // registro en bitacora
+                $acceso_sistema = $this->acceso_sistema_model->get_acceso_sistema($id_acceso_sistema);
+                $accion = "eliminó";
+                $entidad = 'acceso_sistema';
+                $valor = $acceso_sistema['id_rol'] . " " . $acceso_sistema['cod_opcion_sistema'];
+                $this->fn_sis->registro_bitacora($accion, $entidad, $valor);
 
-            return redirect()->to(site_url("acceso_sistema"));
+                // eliminado
+                $this->acceso_sistema_model->delete($id_acceso_sistema);
+                return redirect()->to($url_actual);
+
+            } else {
+                return redirect()->to(site_url("acceso_sistema"));
+            }
         } else {
             return redirect()->to(site_url("login"));
         }
