+++ app/Controllers/Acceso_sistema_usuario.php
@@ -34,22 +34,31 @@ class Acceso_sistema_usuario extends BaseController
         }
     }
 
-    public function eliminar($id_acceso_sistema_usuario)
+    public function eliminar()
     {
         if ($this->session->logueado) {
 
-            // registro en bitacora
-            $acceso_sistema_usuario = $this->acceso_sistema_usuario_model->get_acceso_sistema_usuario($id_acceso_sistema_usuario);
-            $id_usuario = $acceso_sistema_usuario['id_usuario'];
-            $accion = "eliminó";
-            $entidad = 'acceso_sistema_usuario';
-            $valor = $acceso_sistema_usuario['id_usuario'] . " " . $acceso_sistema_usuario['cod_opcion_sistema'];
-            $this->fn_sis->registro_bitacora($accion, $entidad, $valor);
+            $acceso_sistema_usuario = $this->request->getPost();
+            if ($acceso_sistema_usuario) {
+                $id_acceso_sistema_usuario = $acceso_sistema_usuario['id_acceso_sistema_usuario'];
+                $url_actual = $acceso_sistema_usuario['url_actual'];
 
-            // eliminado
-            $this->acceso_sistema_usuario_model->delete($id_acceso_sistema_usuario);
+                // registro en bitacora
+                $acceso_sistema_usuario = $this->acceso_sistema_usuario_model->get_acceso_sistema_usuario($id_acceso_sistema_usuario);
+                $id_usuario = $acceso_sistema_usuario['id_usuario'];
+                $accion = "eliminó";
+                $entidad = 'acceso_sistema_usuario';
+                $valor = $acceso_sistema_usuario['id_usuario'] . " " . $acceso_sistema_usuario['cod_opcion_sistema'];
+                $this->fn_sis->registro_bitacora($accion, $entidad, $valor);
 
-            return redirect()->to(site_url("usuario/detalle/") . $id_usuario );
+                // eliminado
+                $this->acceso_sistema_usuario_model->delete($id_acceso_sistema_usuario);
+
+                return redirect()->to(site_url("usuario/detalle/") . $id_usuario );
+
+            } else {
+                return redirect()->to(site_url("usuario/detalle/") . $id_usuario );
+            }
         } else {
             return redirect()->to(site_url("login"));
         }
